Refactored TeamMangement file into multiple component files
